FUSETOOLS-1585 - Avoid having two times the same WizardPage
Before the pages on the wizard are added twice in
org.fusesource.ide.camel.editor.globalconfiguration.beans.BeanConfigUtil.handleNewClassWizard(IProject,
Shell) , there is a addPages() called which is also called when
wizard.open() is called which was causing that nothing happened when
clicking "Next"
Now, it is called a single time. No next button is displayed.