FUSETOOLS-2571 FUSETOOLS-2572 avoid loading 3 catalogs when a single one is needed
- deactivate classpath listener during creation of the project, it
avoids to load the catalog for the camel version available in the
template source
- use latest loaded Camel Catalog when no project is specified, Graphiti
seems to be project-agnostic in the ImageProvider and thus the project
cannot be load. Previously, it was loading the "default" Camel catalog.
now it is loading the latest loaded catalog (which is often the right
one by side-effect) and if none loaded use the "default as before
- improved progressMonitor in several places to report progress on
catalog loading

Signed-off-by: Aurélien Pupier <apupier@redhat.com>