FUSETOOLS-2997 - force refresh of text field
when the browse button SelctioNAdapter is setting modifying the model,
forcing the Text widget to resynchronize with model. it was supposed to
be done with the notifyListeners but when the "previous-previous" value
was similar the refresh doesn't occur. in theory, the databinding should
take care of it and even the notifyListerners should not have to be
called.

Signed-off-by: Aurélien Pupier <apupier@redhat.com>