FUSETOOLS-3243 - respect usage of productized group id or not
this is providing a fix with minimal change.
The current behavior is quite strange as it is overriding what the user
has provided in the Main tab of the launch configuration every time. it
would be better to provide warning in case the value is suspected to be
not correct instead of silently overriding it

Signed-off-by: Aurélien Pupier <apupier@redhat.com>