FUSETOOLS-3244 - wait project has been built
now that the project is a Maven one, we need to wait the end of the
build.
Another point is that now the version is cached automatically during the
first build of the Maven project which is forcing to not use the cache
in this test. The case when it will happen in real-life should be very
rare.

Signed-off-by: Aurélien Pupier <apupier@redhat.com>