FUSETOOLS-3244 - ensure project are fully imported and configured
in run mode, the test was always failing, with some differen terrors,
but usually because it is trying to use a Java 12 JRE although thanks to
maven pom config set to compiler level 1.8, it should be Java 1.8.
When setting breaking point it was working.

it was a flaky test before, so high chance that the same problem was
occurring, just less often.

set a wait of job in the test code. it would be nice to improve the
production code but the risk is too high to hit a deadlocK, especially
so close to the deadline delivery.

Chances that users hits the issue is low with normally sized projects.

Signed-off-by: Aurélien Pupier <apupier@redhat.com>