FUSETOOLS-3254 - avoid setting back the cursor to beginning when typing
the trick is to avoid to reset value from model to SWT Widget when it is
already the same. this problem is coming by the usage of Listeners
instead of Databinding. It would be better to use Databinding but
requires an important rewrite that we don't have time to do given
current priorities.

Signed-off-by: Aurélien Pupier <apupier@redhat.com>