FUSETOOLS-3248 - reactivate a subpart of NewDataFormatWizardIT tests
not found a way to have tests playing faster in a reliable way fo rnow
so reactivating only a subpart which is covering some specific cases and
the generic one. it won't detect if a dataformat from the catalog is
added with wrong values but this part should be handled by Camel itself
anyway. Also don't detect in case of specific case, this part stays a
problem onb our side.

Signed-off-by: Aurélien Pupier <apupier@redhat.com>