FUSETOOLS2-49 - Add sonar.organization to default
Since we are one of the early adopters of SonarCloud, our project still
belongs to the "Attic" organization - which was originally the only
organization on SonarCloud when the service started. Because of this,
until now, we did not have to specify the "sonar.organization" property
on our project since it was defaulting to the "Attic" org. For
consistency reasons, SonarSource are going to make this property
mandatory, which means that we have to set it to "default".

Signed-off-by: Aurélien Pupier <apupier@redhat.com>