ENTESB-14075 - stop testing templates with latest version
- Fuse runtime is not providing anymore valid artifacts on public
repositories for that and
doesn't want to set it back.
- this is just removal of the test, the problem still exists for
end-users. A cleaner way would be to remove mechanic to look to online
versions, which would remove a great feature for testing against
unreleased version.
- an intermediate way could be to configure other Maven repositories
hidden behind Red Hat firewall (but it requires more work and will
complexify the build)

Signed-off-by: Aurélien Pupier <apupier@redhat.com>