FUSETOOLS2-3384 - provide EAP templates Java 11 following supported patterns
- Upstream doesn't contain anymore the simple CDI spring example. Only
the Java one. It doesn't contain tests anymore. (see
https://github.com/wildfly-extras/wildfly-camel-examples )
- converted from java to xml spring based to benefit from Tooling UI
even if upstream only the Java version is provided
- there are some versions for wildfly inside the pom which is not
managed by the bom. To be clean we would need to add yet another file
version and replacement BUT it wasn't done in previous versions and
there was never an issue raised so I think we can continue this way

Signed-off-by: Aurélien Pupier <apupier@redhat.com>