FUSETOOLS2-3637 - fix Data Transformation Editor opening when build automatically is disabled
- "build automatically is disabled" might be the default with Eclipse
2022-03
- calling an incremental build on the project before we need to access
the generated java classes. As it is incremental, it will take no time
anyway if users have automatic build activated. For the others, it will
take some time but at least it will work.
- spotted that monitors are not correctly defined but do not want to
touch too many things in this part of the code which is very old and
which concerns Data Transformation which is no more recommended for a
long time.

Signed-off-by: Aurélien Pupier <apupier@redhat.com>