FUSETOOLS-3649 - embed jaxb-xjc
it was previously provided by org.jboss.tools.common.jaxb. there is no
viable replacement available through p2. So embed it and export packages
so other bundles can see it.

must move to 2.3.3 version isntead of 2.2.6 to be compatible with
tjakarta.xml.bind provided in the Target Platform.

it is forcing also to import a lot of other packages, that we can detect
as needed at runtime only... so if test coverage is not good enough it
might break. Maybe need to investigate to use a require bundle on
com.sun.xml.bind and jakarta.xml.bind instead

Last but not least, the jsonschema pojo which is in core feature also
requires the import package. Which is making a not nice cyclic
dependency. But well usually everything is installed at once.

Signed-off-by: Aurélien Pupier <apupier@redhat.com>